+++ middleware.go
@@ -97,8 +97,6 @@ func (r *oauthProxy) metricsMiddleware() echo.MiddlewareFunc {
 		},
 		[]string{"code", "method"},
 	)
-
-	// step: register the metric with prometheus
 	prometheus.MustRegisterOrGet(statusMetrics)
 
 	return func(next echo.HandlerFunc) echo.HandlerFunc {
@@ -118,26 +116,20 @@ func (r *oauthProxy) authenticationMiddleware(resource *Resource) echo.Middlewar
 			// step: grab the user identity from the request
 			user, err := r.getIdentity(cx.Request())
 			if err != nil {
-				log.WithFields(log.Fields{
-					"error": err.Error(),
-				}).Errorf("no session found in request, redirecting for authorization")
-
+				log.WithFields(log.Fields{"error": err.Error()}).Errorf("no session found in request, redirecting for authorization")
 				return r.redirectToAuthorization(cx)
 			}
-			// step: inject the user into the context
 			cx.Set(userContextName, user)
 
 			// step: skip if we are running skip-token-verification
 			if r.config.SkipTokenVerification {
 				log.Warnf("skip token verification enabled, skipping verification - TESTING ONLY")
-
 				if user.isExpired() {
 					log.WithFields(log.Fields{
 						"client_ip":  clientIP,
 						"username":   user.name,
 						"expired_on": user.expiresAt.String(),
 					}).Errorf("the session has expired and verification switch off")
-
 					return r.redirectToAuthorization(cx)
 				}
 			} else {
@@ -150,7 +142,6 @@ func (r *oauthProxy) authenticationMiddleware(resource *Resource) echo.Middlewar
 							"client_ip": clientIP,
 							"error":     err.Error(),
 						}).Errorf("access token failed verification")
-
 						return r.accessForbidden(cx)
 					}
 
@@ -161,7 +152,6 @@ func (r *oauthProxy) authenticationMiddleware(resource *Resource) echo.Middlewar
 							"email":      user.name,
 							"expired_on": user.expiresAt.String(),
 						}).Errorf("session expired and access token refreshing is disabled")
-
 						return r.redirectToAuthorization(cx)
 					}
 
@@ -178,7 +168,6 @@ func (r *oauthProxy) authenticationMiddleware(resource *Resource) echo.Middlewar
 							"email":     user.email,
 							"error":     err.Error(),
 						}).Errorf("unable to find a refresh token for user")
-
 						return r.redirectToAuthorization(cx)
 					}
 
@@ -196,7 +185,6 @@ func (r *oauthProxy) authenticationMiddleware(resource *Resource) echo.Middlewar
 						default:
 							log.WithFields(log.Fields{"error": err.Error()}).Errorf("failed to refresh the access token")
 						}
-
 						return r.redirectToAuthorization(cx)
 					}
 					// get the expiration of the new access token
@@ -223,9 +211,8 @@ func (r *oauthProxy) authenticationMiddleware(resource *Resource) echo.Middlewar
 							}
 						}(user.token, token, refresh)
 					}
-					// step: update the with the new access token
+					// update the with the new access token and inject into the context
 					user.token = token
-					// step: inject the user into the context
 					cx.Set(userContextName, user)
 				}
 			}
@@ -247,18 +234,6 @@ func (r *oauthProxy) admissionMiddleware(resource *Resource) echo.MiddlewareFunc
 				return nil
 			}
 			user := cx.Get(userContextName).(*userContext)
-
-			// step: check the audience for the token is us
-			if r.config.ClientID != "" && !user.isAudience(r.config.ClientID) {
-				log.WithFields(log.Fields{
-					"client_id":  r.config.ClientID,
-					"email":      user.email,
-					"expired_on": user.expiresAt.String(),
-					"issuer":     user.audience,
-				}).Warnf("access token audience is not us, redirecting back for authentication")
-
-				return r.accessForbidden(cx)
-			}
 
 			// step: we need to check the roles
 			if roles := len(resource.Roles); roles > 0 {
@@ -336,8 +311,6 @@ func (r *oauthProxy) headersMiddleware(custom []string) echo.MiddlewareFunc {
 
 	return func(next echo.HandlerFunc) echo.HandlerFunc {
 		return func(cx echo.Context) error {
-
-			// step: retrieve the user context if any
 			if user := cx.Get(userContextName); user != nil {
 				id := user.(*userContext)
 				cx.Request().Header.Set("X-Auth-Email", id.email)
@@ -352,7 +325,6 @@ func (r *oauthProxy) headersMiddleware(custom []string) echo.MiddlewareFunc {
 				if r.config.EnableAuthorizationHeader {
 					cx.Request().Header.Set("Authorization", fmt.Sprintf("Bearer %s", id.token.Encode()))
 				}
-
 				// step: inject any custom claims
 				for claim, header := range customClaims {
 					if claim, found := id.claims[claim]; found {